+++ components/organisms/Header/index.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useRef} from 'react'
 import Link from 'next/link'
 import {shrinkHeader} from '@/lib/functions'
-import Navigation from '@/components/molecules/Navigation'
+import Navigation from '@/components/organisms/Navigation'
 
 const Logo: React.FC = () => {
   return (
