+++ components/organisms/Hero/Hero.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Hero from './Hero'
+import Hero from '.'
 
 /**
  * The following is a story.
